Create Multiplayer Mechanic for update Board per client

+++ JengaTetris/Assets/Scripts/gamenet/event/GamenetEvent.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using UnityEngine.Events;
 using UnityEngine.Networking;
 
-public class GamenetEvent : UnityEvent<NetworkConnection, short>
+public class GamenetEvent : UnityEvent<NetworkConnection, int>
 {
 
 }
